Fix size of uint for next hop group ID
Summary:
r-barnes pointed out that the loop guard is comparing uint32_t and uint64 max, hence it will always return true.

Fix by explicitly defining the type of uint64_t.

Reviewed By: r-barnes

Differential Revision:
D89504816

Privacy Context Container: L1297311

fbshipit-source-id: 8740ee9a005e045db79f41175916a08835608e5f

+++ fboss/agent/EcmpResourceManager.h
@@ -407,7 +407,7 @@ class EcmpResourceManager {
       const std::shared_ptr<Route<AddrT>>& removed,
       bool isUpdate,
       InputOutputState* inOutState);
-  static uint32_t constexpr kMinNextHopGroupId = 1;
+  static uint64_t constexpr kMinNextHopGroupId = 1;
   NextHopGroupId findCachedOrNewIdForNhops(
       const RouteNextHopSet& nhops,
       const InputOutputState& inOutState) const;
